DestinationHint and Angel bugfix 2 BoundingBox improved

+++ trunk/controltheland/src/main/java/com/btxtech/game/services/bot/impl/BotItemContainer.java
@@ -85,11 +85,11 @@ public class BotItemContainer {
     }
 
     public BotSyncBaseItem getFirstIdleAttacker(SyncBaseItem target) {
-        int distance = Integer.MAX_VALUE;
+        double distance = Integer.MAX_VALUE;
         BotSyncBaseItem attacker = null;
         for (BotSyncBaseItem botSyncBaseItem : botItems.values()) {
             if (botSyncBaseItem.isIdle() && botSyncBaseItem.isAbleToAttack(target.getBaseItemType())) {
-                int dist = botSyncBaseItem.getDistanceTo(target);
+                double dist = botSyncBaseItem.getDistanceTo(target);
                 if (dist < distance) {
                     distance = dist;
                     attacker = botSyncBaseItem;
